llm_resume_grader/requirements.txt
openai>=1.12
pyyaml
python-dotenv
pytest
